合并 jade 的 Provider
