+++ src/main/java/com/epimorphismmc/monazite/integration/jade/MonaziteJadePlugin.java
@@ -1,8 +1,6 @@
 package com.epimorphismmc.monazite.integration.jade;
 
-import com.epimorphismmc.monazite.config.MonaziteConfigHolder;
-import com.epimorphismmc.monazite.integration.jade.provider.RecipeFluidOutputInfoProvider;
-import com.epimorphismmc.monazite.integration.jade.provider.RecipeItemOutputInfoProvider;
+import com.epimorphismmc.monazite.integration.jade.provider.RecipeOutputInfoProvider;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import snownee.jade.api.IWailaClientRegistration;
@@ -15,24 +13,12 @@ public class MonaziteJadePlugin implements IWailaPlugin {
 
     @Override
     public void register(IWailaCommonRegistration registration) {
-        if (MonaziteConfigHolder.INSTANCE.topInformation.displayItemOutputs) {
-            registration.registerBlockDataProvider(RecipeItemOutputInfoProvider.INSTANCE, BlockEntity.class);
-        }
-
-        if (MonaziteConfigHolder.INSTANCE.topInformation.displayFluidOutputs) {
-            registration.registerBlockDataProvider(RecipeFluidOutputInfoProvider.INSTANCE, BlockEntity.class);
-        }
+        registration.registerBlockDataProvider(RecipeOutputInfoProvider.INSTANCE, BlockEntity.class);
     }
 
     @Override
     public void registerClient(IWailaClientRegistration registration) {
-        if (MonaziteConfigHolder.INSTANCE.topInformation.displayItemOutputs) {
-            registration.registerBlockComponent(RecipeItemOutputInfoProvider.INSTANCE, Block.class);
-        }
-
-        if (MonaziteConfigHolder.INSTANCE.topInformation.displayFluidOutputs) {
-            registration.registerBlockComponent(RecipeFluidOutputInfoProvider.INSTANCE, Block.class);
-        }
+        registration.registerBlockComponent(RecipeOutputInfoProvider.INSTANCE, Block.class);
     }
